+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Updated [README.md]
+- Added [Join2-Object] `-OnlyIfInTransaction` switch
 
 ## [6.3.0] - 2023-10-14